Skip build for legacy topo aware scheduler.
Skip build for legacy topology aware scheduler through build tag.
