+++ pkg/algorithm/topology_aware_scheduler.go
@@ -20,6 +20,8 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE
 
+// +build ignore
+
 package algorithm
 
 import (
@@ -438,27 +440,6 @@ func removePickedLeafCells(leafCells CellList, indices []int32) CellList {
 		leafCells[i] = nil
 	}
 	return leafCells[:len(leafCells)-len(indices)]
-}
-
-// findLCA finds the lowest common ancestor of two cells (nil if they have no LCA).
-func findLCA(lower Cell, higher Cell) Cell {
-	for lower.GetLevel() < higher.GetLevel() {
-		if lower.GetParent() == nil {
-			return nil
-		}
-		lower = lower.GetParent()
-	}
-	if CellEqual(lower, higher) {
-		return lower
-	}
-	for !CellEqual(lower.GetParent(), higher.GetParent()) {
-		if lower.GetParent() == nil || higher.GetParent() == nil {
-			return nil
-		}
-		lower = lower.GetParent()
-		higher = higher.GetParent()
-	}
-	return lower.GetParent()
 }
 
 // getLeafCellsFromNode collects free leaf cells and preemptible leaf cells according to the priority.
